combine Fig. 6, 7
- combine Fig. 6, 7 into a single figure with 3 panels
- remove old Fig. 6, 7
- rename as Fig. 8
